requirements.txt
Django==1.3
PIL==1.1.7
Whoosh==2.3.2
distribute==0.6.24
django-haystack==1.2.7
django-registration==0.8
psycopg2==2.4.5
wsgiref==0.1.2
